periods: Introduce DataTable component in BudgetingPeriodsList
